+++ graylog2-web-interface/src/views/logic/fieldactions/ChangeFieldType/ChangeFieldTypeModal.tsx
@@ -14,41 +14,21 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-import React, { useMemo, useCallback, useState, useContext, useEffect } from 'react';
+import React, { useMemo, useCallback, useState } from 'react';
 import styled from 'styled-components';
 
 import {
-  PaginatedList,
-  SearchForm,
-  Spinner,
-  NoSearchResult,
-  NoEntitiesExist,
-  EntityDataTable,
   Icon,
   Select,
 } from 'components/common';
-import { Button, BootstrapModalForm, Alert, Input } from 'components/bootstrap';
-import { DEFAULT_LAYOUT, ENTITY_TABLE_ID } from 'views/logic/fieldactions/ChangeFieldType/Constants';
-import useTableLayout from 'components/common/EntityDataTable/hooks/useTableLayout';
-import useFiledTypeUsages from 'views/logic/fieldactions/ChangeFieldType/hooks/useFiledTypeUsages';
-import useUpdateUserLayoutPreferences from 'components/common/EntityDataTable/hooks/useUpdateUserLayoutPreferences';
-import type { Sort } from 'stores/PaginationTypes';
-import type { FieldTypeUsage, ChangeFieldTypeFormValues } from 'views/logic/fieldactions/ChangeFieldType/types';
-import useColumnRenderers from 'views/logic/fieldactions/ChangeFieldType/hooks/useColumnRenderers';
-import QueryHelper from 'components/common/QueryHelper';
-import BulkActionsDropdown from 'components/common/EntityDataTable/BulkActionsDropdown';
+import { BootstrapModalForm, Alert, Input } from 'components/bootstrap';
+import type { ChangeFieldTypeFormValues } from 'views/logic/fieldactions/ChangeFieldType/types';
 import useFiledTypes from 'views/logic/fieldactions/ChangeFieldType/hooks/useFieldTypes';
-import { ActionContext } from 'views/logic/ActionContext';
-import useCurrentQuery from 'views/logic/queries/useCurrentQuery';
-import { filtersToStreamSet } from 'views/logic/queries/Query';
+import IndexSetsTable from 'views/logic/fieldactions/ChangeFieldType/IndexSetsTable';
 
 const StyledSelect = styled(Select)`
   width: 400px;
   margin-bottom: 20px;
-`;
-
-const Container = styled.div`
-  margin-top: 20px;
 `;
 
 type Props = {
@@ -57,17 +37,7 @@ type Props = {
   onSubmit: (formValues: ChangeFieldTypeFormValues) => void,
   onClose: () => void }
 
-const renderBulkActions = (
-  selectedDashboardIds: Array<string>,
-  setSelectedDashboardIds: (streamIds: Array<string>) => void,
-) => (
-  <BulkActionsDropdown selectedEntities={selectedDashboardIds} setSelectedEntities={setSelectedDashboardIds} />
-);
-
 const ChangeFieldTypeModal = ({ show, onClose, onSubmit, field }: Props) => {
-  const [query, setQuery] = useState('');
-  const [showDetails, setShowDetails] = useState(false);
-  const [activePage, setActivePage] = useState(1);
   const [rotated, setRotated] = useState(false);
   const [newFieldType, setNewFieldType] = useState(null);
   const { data: { fieldTypes }, isLoading: isOptionsLoading } = useFiledTypes();
@@ -75,84 +45,13 @@ const ChangeFieldTypeModal = ({ show, onClose, onSubmit, field }: Props) => {
     id,
     label,
   })), [fieldTypes]);
-  const { widget, message } = useContext(ActionContext);
-  const currentQuery = useCurrentQuery();
-  const currentStreams = useMemo(() => message?.fields?.streams ?? widget?.streams ?? filtersToStreamSet(currentQuery.filter).toJS(), [message?.fields?.streams, currentQuery.filter, widget?.streams]);
-
-  const { layoutConfig, isInitialLoading: isLoadingLayoutPreferences } = useTableLayout({
-    entityTableId: ENTITY_TABLE_ID,
-    defaultPageSize: DEFAULT_LAYOUT.pageSize,
-    defaultDisplayedAttributes: DEFAULT_LAYOUT.displayedColumns,
-    defaultSort: DEFAULT_LAYOUT.sort,
-  });
-
-  const searchParams = useMemo(() => ({
-    query,
-    page: activePage,
-    pageSize: layoutConfig.pageSize,
-    sort: layoutConfig.sort,
-  }), [activePage, layoutConfig.pageSize, layoutConfig.sort, query]);
-  const { data: { list, attributes, pagination }, isFirsLoaded } = useFiledTypeUsages({ field, streams: currentStreams }, searchParams, { enabled: !isLoadingLayoutPreferences && !!currentStreams });
-  const initialSelection = useMemo(() => list.map(({ id }) => id), [list]);
 
   const [indexSetSelection, setIndexSetSelection] = useState<Array<string>>();
-  const { mutate: updateTableLayout } = useUpdateUserLayoutPreferences(ENTITY_TABLE_ID);
-
-  useEffect(() => {
-    setIndexSetSelection(initialSelection);
-  }, [initialSelection]);
-
-  const onPageChange = useCallback(
-    (newPage: number, newPageSize: number) => {
-      if (newPage) {
-        setActivePage(newPage);
-      }
-
-      if (newPageSize) {
-        updateTableLayout({ perPage: newPageSize });
-      }
-    }, [updateTableLayout],
-  );
-
-  const onPageSizeChange = useCallback((newPageSize: number) => {
-    setActivePage(1);
-    updateTableLayout({ perPage: newPageSize });
-  }, [updateTableLayout]);
-
-  const onSortChange = useCallback((newSort: Sort) => {
-    setActivePage(1);
-    updateTableLayout({ sort: newSort });
-  }, [updateTableLayout]);
-
-  const onSearch = useCallback((newQuery: string) => {
-    setActivePage(1);
-    setQuery(newQuery);
-  }, []);
-
-  const onResetSearch = useCallback(() => onSearch(''), [onSearch]);
-
-  const onColumnsChange = useCallback((displayedAttributes: Array<string>) => {
-    updateTableLayout({ displayedAttributes });
-  }, [updateTableLayout]);
-
-  const columnRenderers = useColumnRenderers(fieldTypes);
-
-  const onChangeSelection = useCallback((newSelection) => {
-    setIndexSetSelection(newSelection);
-  }, []);
-
-  const toggleDetailsOpen = useCallback(() => {
-    setShowDetails((cur) => !cur);
-  }, []);
 
   const _onSubmit = useCallback((e) => {
     e.preventDefault();
     onSubmit({ indexSetSelection, newFieldType, rotated });
   }, [indexSetSelection, newFieldType, onSubmit, rotated]);
-
-  if (isLoadingLayoutPreferences || !isFirsLoaded || isOptionsLoading) {
-    return <Spinner />;
-  }
 
   return (
     <BootstrapModalForm title={`Change ${field} field type`}
@@ -172,61 +71,13 @@ const ChangeFieldTypeModal = ({ show, onClose, onSubmit, field }: Props) => {
                       value={newFieldType}
                       onChange={(value) => setNewFieldType(value)}
                       placeholder="Select field type"
+                      disabled={isOptionsLoading}
                       required />
         <Alert bsStyle="info">
           <Icon name="info-circle" />&nbsp;
           By default the type will be changed in all possible indexes. But you can choose in which index sets you would like to make the change
         </Alert>
-        <Button bsStyle="link" className="btn-text" bsSize="xsmall" onClick={toggleDetailsOpen}>
-          <Icon name={`caret-${showDetails ? 'down' : 'right'}`} />&nbsp;
-          {showDetails ? 'Hide index sets' : 'Show index sets'}
-        </Button>
-        {
-        showDetails && (
-          <Container>
-            <PaginatedList onChange={onPageChange}
-                           totalItems={pagination?.total}
-                           pageSize={layoutConfig.pageSize}
-                           activePage={activePage}
-                           showPageSizeSelect={false}
-                           useQueryParameter={false}>
-              <div style={{ marginBottom: 5 }}>
-                <SearchForm onSearch={onSearch}
-                            queryHelpComponent={<QueryHelper entityName="dashboard" commonFields={['id', 'title', 'description', 'summary']} />}
-                            onReset={onResetSearch}
-                            query={query}
-                            topMargin={0} />
-              </div>
-              {!list?.length && !query && (
-              <NoEntitiesExist>
-                No index sets have been found.
-              </NoEntitiesExist>
-              )}
-              {!list?.length && query && (
-              <NoSearchResult>No index sets have been found.</NoSearchResult>
-              )}
-              {list.length && (
-              <EntityDataTable<FieldTypeUsage> activeSort={layoutConfig.sort}
-                                               bulkSelection={{
-                                                 onChangeSelection,
-                                                 initialSelection,
-                                                 actions: renderBulkActions,
-                                               }}
-                                               columnDefinitions={attributes}
-                                               columnRenderers={columnRenderers}
-                                               columnsOrder={DEFAULT_LAYOUT.columnsOrder}
-                                               data={list}
-                                               onColumnsChange={onColumnsChange}
-                                               onPageSizeChange={onPageSizeChange}
-                                               onSortChange={onSortChange}
-                                               pageSize={layoutConfig.pageSize}
-                                               visibleColumns={layoutConfig.displayedAttributes} />
-              )}
-            </PaginatedList>
-          </Container>
-        )
-      }
-
+        <IndexSetsTable field={field} setIndexSetSelection={setIndexSetSelection} fieldTypes={fieldTypes} />
         <Input type="checkbox"
                id="rotate"
                name="rotate"
